+++ spec/accessibility/all_projects_spec.rb
@@ -38,17 +38,17 @@ RSpec.feature "All projects", driver: :headless_firefox, accessibility: true do
     check_accessibility(page)
   end
 
-  scenario "> Opening > Confirmed conversion date" do
+  scenario "> By month > Confirmed conversion date" do
     project = create(:conversion_project, assigned_to: user, urn: 123434, conversion_date: Date.today.next_month.at_beginning_of_month, conversion_date_provisional: false)
 
     visit confirmed_all_by_month_projects_path
 
     expect(page).to have_content(project.urn)
-    expect(page).to have_link("Opening")
+    expect(page).to have_link("Proceeding as planned")
     check_accessibility(page)
   end
 
-  scenario "> Opening > Revised conversion date" do
+  scenario "> By month > Revised conversion date" do
     project = create(:conversion_project, assigned_to: user, urn: 123434, conversion_date_provisional: false)
     create(:date_history, project: project, previous_date: Date.today.next_month.at_beginning_of_month)
     create(:date_history, project: project, previous_date: Date.today.next_month.at_beginning_of_month)
@@ -56,7 +56,7 @@ RSpec.feature "All projects", driver: :headless_firefox, accessibility: true do
     visit revised_all_by_month_projects_path
 
     expect(page).to have_content(project.urn)
-    expect(page).to have_link("Opening")
+    expect(page).to have_link("Proceeding as planned")
     check_accessibility(page)
   end
 end